django/requirements.txt
django==5.2.3
django-environ==0.10.0
django-cors-headers==4.3.1
django-random-id-model==0.1.1
django-querycount==0.8.3